requirements.txt
django
requests
djangorestframework
django-cors-headers
python-dotenv
django-filter
drf-spectacular
drf-spectacular-sidecar
djangorestframework-simplejwt